+++ web/src/views/Game/index.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import TopNav from '../../layout/TopNav'
 import Loading from '../../components/Loading'
+import { fetchGame } from "../../api"
 
 import Team from './Team'
 import Points from './Points'
@@ -18,9 +19,10 @@ export default class Game extends Component {
   componentDidMount () {
     const gameId = this.props.match.params.gameId
 
-    fetch(`/api/games/${gameId}`)
-      .then(response => response.json())
-      .then(game => { this.setState({loading: false, game: game}) })
+    return (async () => {
+      const game = await fetchGame(gameId)
+      this.setState({game, loading: false})
+    })()
   }
 
   renderMain () {
